poser templates now have a base class
also thread_register->register_member_thread, its now a static method of the poser template, should only be used to register member functions when the template is inherited
