+++ virtualreality/templates/templates.py
@@ -4,9 +4,10 @@ import numbers
 import warnings
 
 from ..util import utilz as u
+from .template_base import *
 from .poses import *
 
-class PoserTemplate:
+class PoserTemplate(PoserTemplateBase):
     """
     Poser base class.
 
@@ -40,7 +41,7 @@ class PoserTemplate:
     Example:
         class MyPoser(PoserTemplate):
             def __init__(self, *args, **kwargs):
-                super().__init(**kwargs)
+                super().__init__(**kwargs)
 
             @thread_register(0.5)
             async def example_thread(self):
@@ -70,46 +71,11 @@ class PoserTemplate:
         :param send_delay: sleep delay for the self.send thread(in seconds)
         :param recv_delay: sleep delay for the self.recv thread(in seconds)
         """
-        self.addr = addr
-        self.port = port
-
-        self._coro_name_exceptions = ["main"]
-
-        self.coro_list = [
-            method_name
-            for method_name in dir(self)
-            if callable(getattr(self, method_name))
-            and method_name[0] != "_"
-            and method_name not in self._coro_name_exceptions
-        ]
-
-        self.coro_keep_alive = {
-            "close": [True, 0.1],
-            "send": [True, send_delay],
-            "recv": [True, recv_delay],
-        }
-        self.last_read = ""
-        self.id_message = 'holla'
+        super().__init__(addr=addr, port=port, send_delay=send_delay, recv_delay=recv_delay)
 
         self.pose = Pose()
         self.pose_controller_r = ControllerState(pose=(0.5, 1, -1, 1, 0, 0, 0))
         self.pose_controller_l = ControllerState(pose=(0.5, 1.1, -1, 1, 0, 0, 0))
-
-    async def _socket_init(self):
-        """
-        Connect to the server using self.addr and self.port and send the id message.
-
-        Also store socket reader and writer in self.reader and self.writer.
-        It is not recommended you override this method
-
-        send id message
-        more on id messages: https://github.com/okawo80085/hobo_vr/wiki/server-id-messages
-        """
-        print(f'connecting to the server at "{self.addr}:{self.port}"...')
-        # connect to the server
-        self.reader, self.writer = await asyncio.open_connection(self.addr, self.port, loop=asyncio.get_event_loop())
-        # send poser id message
-        self.writer.write(u.format_str_for_write(self.id_message))
 
     async def send(self):
         """Send all poses thread."""
@@ -131,102 +97,6 @@ class PoserTemplate:
                 print(f"send failed: {e}")
                 self.coro_keep_alive["send"][0] = False
                 break
-
-    async def recv(self):
-        """Receive messages thread."""
-        while self.coro_keep_alive["recv"][0]:
-            try:
-                data = await u.read(self.reader)
-                self.last_read = data
-
-                await asyncio.sleep(self.coro_keep_alive["recv"][1])
-            except Exception as e:
-                print(f"recv failed: {e}")
-                self.coro_keep_alive["recv"][0] = False
-                break
-
-    async def close(self):
-        """Await close thread, press "q" to kill all registered threads."""
-        try:
-            import keyboard
-
-            while self.coro_keep_alive["close"][0]:
-                if keyboard.is_pressed("q"):
-                    break
-
-                await asyncio.sleep(self.coro_keep_alive["close"][1])
-
-        except ImportError as e:
-            print(f"close: failed to import keyboard, poser will close in 10 seconds: {e}")
-            await asyncio.sleep(10)
-
-        print("closing...")
-        self.coro_keep_alive["close"][0] = False
-        for key in self.coro_keep_alive:
-            if self.coro_keep_alive[key][0]:
-                self.coro_keep_alive[key][0] = False
-                print(f"{key} stop sent...")
-            else:
-                print(f"{key} already stopped")
-
-        await asyncio.sleep(0.5)
-
-        try:
-            self.writer.write(u.format_str_for_write("CLOSE"))
-            self.writer.close()
-            await self.writer.wait_closed()
-
-        except Exception as e:
-            print (f'failed to close connection: {e}')
-
-        print("finished")
-
-    async def main(self):
-        """
-        Run the main async method.
-
-        Gathers all recognized threads and runs them.
-        It is not recommended you override this method.
-        """
-        await self._socket_init()
-
-        await asyncio.gather(
-            *[getattr(self, coro_name)() for coro_name in self.coro_list if coro_name not in self._coro_name_exceptions]
-        )
-
-
-def thread_register(sleepDelay, runInDefaultExecutor=False):
-    """
-    Register thread for PoserTemplate base class.
-
-    sleepDelay - sleep delay in seconds
-    runInDefaultExecutor - bool, set True if you want the function to be executed in asyncio's default pool executor
-    """
-
-    def _thread_reg(func):
-        def _thread_reg_wrapper(self, *args, **kwargs):
-
-            if not asyncio.iscoroutinefunction(func) and not runInDefaultExecutor:
-                raise ValueError(f"{repr(func)} is not a coroutine function and runInDefaultExecutor is set to False")
-
-            if func.__name__ not in self.coro_keep_alive and func.__name__ in self.coro_list:
-                self.coro_keep_alive[func.__name__] = [True, sleepDelay]
-
-            else:
-                warnings.warn("thread register ignored, thread already exists", RuntimeWarning)
-
-            if runInDefaultExecutor:
-                loop = asyncio.get_running_loop()
-
-                return loop.run_in_executor(None, func, self, *args, **kwargs)
-
-            return func(self, *args, **kwargs)
-
-        setattr(_thread_reg_wrapper, "__name__", func.__name__)
-
-        return _thread_reg_wrapper
-
-    return _thread_reg
 
 
 class PoserClient(PoserTemplate):
